requirements.txt
azure-functions
google-genai
requests
python-dotenv
PyMuPDF